agent/requirements.txt
psutil
kafka-python